MISRA12: General cleanup to be more compliant to the "Essential Type Model" section

+++ kernel/src/driver/ramdisk.c
@@ -35,7 +35,7 @@ static struct
 
 static ssize_t ramdisk_rw_block(void *buf, size_t blocknum, int doread)
 {
-    ssize_t n = BLOCK_SIZE;
+    size_t n = BLOCK_SIZE;
     size_t off;
 
     off = blocknum * BLOCK_SIZE;
@@ -48,7 +48,7 @@ static ssize_t ramdisk_rw_block(void *buf, size_t blocknum, int doread)
         memcpy(buf, (char *)ramdisk.addr + off, n);
     else
         memcpy((char *)ramdisk.addr + off, buf, n);
-    return n;
+    return (ssize_t)n;
 }
 
 static ssize_t ramdisk_read_block(void *buf, size_t blocknum)
@@ -68,7 +68,7 @@ static char blk[BLOCK_SIZE];
 ssize_t ramdisk_read(void *buf, size_t size, size_t off)
 {
     unsigned int nblk;
-    ssize_t left;
+    size_t left;
     size_t ioff;
     ssize_t n;
     char *ptr;
@@ -80,7 +80,7 @@ ssize_t ramdisk_read(void *buf, size_t size, size_t off)
 
     if ((n = ramdisk_read_block(blk, nblk)) != BLOCK_SIZE)
         return n;
-    n = MIN(BLOCK_SIZE-ioff, left);
+    n = MIN(BLOCK_SIZE - ioff, left);
     memcpy(ptr, &blk[ioff], n);
     left -= n;
     ptr += n;
@@ -96,7 +96,7 @@ ssize_t ramdisk_read(void *buf, size_t size, size_t off)
         nblk++;
     }
 
-    return size-left;
+    return (ssize_t)(size - left);
 }
 
 ssize_t ramdisk_write(const void *buf, size_t size, size_t off)
